Fix the resume feature
